Put builder into top level
