+++ application/src/main/java/org/opentripplanner/osm/model/OsmRelation.java
@@ -2,7 +2,6 @@ package org.opentripplanner.osm.model;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.opentripplanner.osm.OsmProvider;
@@ -11,7 +10,7 @@ public class OsmRelation extends OsmEntity {
 
   private final List<OsmRelationMember> members;
 
-  private OsmRelation(
+  public OsmRelation(
     long id,
     Map<String, String> tags,
     OsmProvider osmProvider,
@@ -21,55 +20,16 @@ public class OsmRelation extends OsmEntity {
     this.members = Collections.unmodifiableList(new ArrayList<>(members));
   }
 
-  public static Builder builder() {
-    return new Builder();
+  public static OsmRelationBuilder builder() {
+    return new OsmRelationBuilder();
   }
 
-  public Builder copy() {
-    Builder builder = new Builder();
-    builder.id = this.id;
-    builder.tags = new HashMap<>(this.getTags());
-    builder.osmProvider = this.getOsmProvider();
-    builder.members = new ArrayList<>(this.members);
-    return builder;
+  public OsmRelationBuilder copy() {
+    return new OsmRelationBuilder(this);
   }
 
   public List<OsmRelationMember> getMembers() {
     return members;
-  }
-
-  public static class Builder {
-
-    private long id;
-    private Map<String, String> tags = new HashMap<>();
-    private OsmProvider osmProvider;
-    private List<OsmRelationMember> members = new ArrayList<>();
-
-    public Builder withId(long id) {
-      this.id = id;
-      return this;
-    }
-
-    public Builder addTag(String key, String value) {
-      if (key != null && value != null) {
-        this.tags.put(key.toLowerCase(), value);
-      }
-      return this;
-    }
-
-    public Builder withOsmProvider(OsmProvider osmProvider) {
-      this.osmProvider = osmProvider;
-      return this;
-    }
-
-    public Builder addMember(OsmRelationMember member) {
-      members.add(member);
-      return this;
-    }
-
-    public OsmRelation build() {
-      return new OsmRelation(id, tags, osmProvider, members);
-    }
   }
 
   public String toString() {
